Just some hellowords about rabbitMQ and node.js
different design patterns to use different exchanges and queues
configurations.

The last example appPubSub include a more efficient mechanisms
to facilitate pub and sub mechanisms. One program launch the
channels and make the bindings with the queues, from now on
the publisher just need and ID for que exchange and the sub
the ID for a queue where receive messages