Send messages through a rabbitMQ instance
just a simple REST interface to manage
 /entities/id/channel/id/topic/pub

to publish messages through an specific channel about an
specific topic.

Internally, the rabbit broker has been configured with
a set of exchanges binded with queues.

In this case the client(REST) just needs to know the channel-id
ant the topic.

By other hand, you could create your listening clients that
should be attachd to the broker with the pre-configured queues.